feat: ONNX classifier module

+++ src/obd_parser.hpp
@@ -6,13 +6,13 @@
 
 struct OBDRecord
 {
-    double speed;
-    double rpm;
-    double throttlePos;
-    double coolantTemp;
-    double fuelLevel;
-    double intakeTemp;
-    double label;
+    float speed;
+    float rpm;
+    float throttlePos;
+    float coolantTemp;
+    float fuelLevel;
+    float intakeTemp;
+    float label;
 };
 
 #pragma pack(pop)
@@ -24,12 +24,12 @@ public:
     static OBDParser create();
     int64_t load(std::filesystem::path filePath);
     OBDRecord getRecord(int64_t pos);
-    std::map<std::string, double> getMapLabel();
+    std::map<std::string, float> getMapLabel();
 
 private:
     OBDParser();
     void initialize();
 
     std::vector<OBDRecord> m_records;
-    std::map<std::string, double> m_mapLabel;
+    std::map<std::string, float> m_mapLabel;
 };
